add more tests for youtube code
